refactor: Re-include the keyword `Override` in the file + function names.
