+++ helpers/channels/deleteChannelPermissionOverride.ts
@@ -1,7 +1,7 @@
 import type { Bot } from "../../bot.ts";
 
 /** Delete the channel permission overwrites for a user or role in this channel. Requires `MANAGE_ROLES` permission. */
-export async function deleteChannelPermissions(bot: Bot, channelId: bigint, overwriteId: bigint): Promise<void> {
+export async function deleteChannelPermissionOverride(bot: Bot, channelId: bigint, overwriteId: bigint): Promise<void> {
   return await bot.rest.runMethod<void>(
     bot.rest,
     "DELETE",
