+++ helpers/channels/editChannel.ts
@@ -2,7 +2,7 @@ import type { Bot } from "../../bot.ts";
 import { Channel } from "../../transformers/channel.ts";
 import { DiscordChannel } from "../../types/discord.ts";
 import { ChannelTypes, VideoQualityModes } from "../../types/shared.ts";
-import { OverwriteReadable } from "./editChannelPermissions.ts";
+import { OverwriteReadable } from "./editChannelPermissionOverrides.ts";
 
 /** Update a channel's settings. Requires the `MANAGE_CHANNELS` permission for the guild. */
 export async function editChannel(
